options: Factor out options_string code from options.c
options.c is a huge, unwieldly file. So remove some
closely related code that has no direct dependencies
with the other functions in the file.

This time it is the functions related to the creation
and testing of the options string.

Acked-by: Razvan Cojocaru <razvanc@mailbox.org>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1773
Message-Id: <20260820124355.23332-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg38563.html
