+++ CMakeLists.txt
@@ -533,6 +533,8 @@ set(SOURCE_FILES
     src/openvpn/options_parse.c
     src/openvpn/options_show.c
     src/openvpn/options_show.h
+    src/openvpn/options_string.c
+    src/openvpn/options_string.h
     src/openvpn/otime.c
     src/openvpn/otime.h
     src/openvpn/ovpn_dco_win.h
